requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.0
frozenlist==1.4.1
greenlet==3.1.1
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.14
langchain-core==0.3.29
langchain-ollama==0.2.2
langchain-text-splitters==0.3.4
langsmith==0.2.10
multidict==6.1.0
numpy==2.2.1
ollama==0.4.5
orjson==3.10.13
packaging==24.2
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.36
srt==3.5.3
starlette==0.38.6
tenacity==9.0.0
tqdm==4.67.1
twilio==9.3.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.21.0
vosk==0.3.45
watchfiles==1.0.3
websockets==13.1
yarl==1.12.1
